requirements-webapp.txt
flask==2.3.3
flask-login==0.6.3
flask-wtf==1.2.1
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
pyotp==2.9.0
qrcode[pil]==7.4.2
requests==2.31.0
pandas==2.1.4
plotly==5.18.0
jinja2==3.1.2
weasyprint==60.2
python-magic==0.4.27
python-dateutil==2.8.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
apscheduler==3.10.4
werkzeug==2.3.7
ansible==8.6.0
pyyaml==6.0.1

